refactor(injection): move tests into per-submodule test modules
The previous split left the entire ~1650-line test module in the
injection.rs facade, while the sibling modules it mirrors
(analysis/semantic/*, analysis/selection/*) each carry their own
#[cfg(test)] mod tests. This finishes the relocation one level down so
each concern owns its tests:

- offset.rs    — #offset! directive parsing tests
- ranges.rs    — included-range math tests (+ make_range helper)
- content.rs   — clean-content / coordinate tests (+ blockquote helper)
- discovery.rs — region discovery + resolver integration tests

injection.rs is now a pure facade (mod declarations + re-exports, no
test module). is_node_within / find_injection_at_position revert to
private now that their tests live beside them; only
parse_offset_directive_for_pattern keeps a cfg(test) re-export, for the
selection-range tests that exercise it through the public path. Pure
structural Tidy: same tests, same assertions, behavior unchanged.

Refs #393
