+++ src/language/injection/content.rs
@@ -162,3 +162,231 @@ pub(crate) fn byte_to_point_anchored(
         },
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::language::injection::{collect_all_injections, compute_included_ranges};
+    use tree_sitter::{Parser, Query, StreamingIterator};
+
+    /// Helper to set up a markdown parser and injection query, parse text,
+    /// and return the first injection's content node byte range and included ranges.
+    fn blockquote_injection_data(
+        text: &str,
+    ) -> (std::ops::Range<usize>, u32, Option<Vec<tree_sitter::Range>>) {
+        let mut parser = Parser::new();
+        let md_language: tree_sitter::Language = tree_sitter_md::LANGUAGE.into();
+        parser
+            .set_language(&md_language)
+            .expect("load markdown grammar");
+
+        let injection_query_str = r#"
+            (fenced_code_block
+              (info_string (language) @injection.language)
+              (code_fence_content) @injection.content)
+        "#;
+        let injection_query =
+            Query::new(&md_language, injection_query_str).expect("valid injection query");
+
+        let tree = parser.parse(text, None).expect("parse markdown");
+        let root = tree.root_node();
+
+        let injections = collect_all_injections(&root, text, Some(&injection_query))
+            .expect("Should find injections");
+        assert!(!injections.is_empty(), "Should find at least one injection");
+
+        let region = &injections[0];
+        let byte_range = region.content_node.byte_range();
+        let included_ranges =
+            compute_included_ranges(&region.content_node, region.include_children);
+
+        // Compute start_column (UTF-16)
+        let byte_column = region.content_node.start_position().column;
+        let line_start_byte = region.content_node.start_byte() - byte_column;
+        let line_prefix = &text[line_start_byte..region.content_node.start_byte()];
+        let start_column = line_prefix.encode_utf16().count() as u32;
+
+        (byte_range, start_column, included_ranges)
+    }
+
+    #[test]
+    fn test_byte_to_point_anchored_matches_full_scan() {
+        // Exhaustive equivalence with the full-prefix scan, covering the
+        // before-anchor fallback, same-row and cross-row targets, and
+        // mid-codepoint clamping (bytes inside あ floor to its start).
+        let text = "abc\ndefあ\nghi";
+        let anchor_byte = 4; // start of "def"
+        let anchor_point = byte_to_point(text, anchor_byte);
+        for byte in 0..=text.len() {
+            assert_eq!(
+                byte_to_point_anchored(text, byte, anchor_byte, anchor_point),
+                byte_to_point(text, byte),
+                "anchored and full scans must agree at byte {byte}"
+            );
+        }
+    }
+
+    #[test]
+    fn test_blockquote_bridge_content_strips_prefixes() {
+        // Verify that extract_clean_content strips blockquote prefixes so that
+        // downstream language servers receive parseable code.
+        let text = "> ```lua\n> local x = 1\n> local y = 2\n> ```\n";
+        let (byte_range, _start_column, included_ranges) = blockquote_injection_data(text);
+
+        let clean = extract_clean_content(text, byte_range, included_ranges.as_deref());
+        assert!(
+            !clean.contains("> "),
+            "Clean content should NOT contain blockquote prefixes: {:?}",
+            clean
+        );
+        assert!(
+            clean.contains("local x = 1"),
+            "Clean content should contain the code: {:?}",
+            clean
+        );
+        assert!(
+            clean.contains("local y = 2"),
+            "Clean content should contain the code: {:?}",
+            clean
+        );
+    }
+
+    #[test]
+    fn extract_clean_content_without_ranges_returns_full_text() {
+        // When called with None for included_ranges, returns the full content slice
+        let text = "hello world";
+        let clean = extract_clean_content(text, 0..text.len(), None);
+        assert_eq!(clean, text);
+    }
+
+    #[test]
+    fn extract_clean_content_with_ranges_strips_prefixes() {
+        // Blockquote case: included_ranges present → only gap content
+        let text = "> ```lua\n> local x = 1\n> local y = 2\n> ```\n";
+        let (byte_range, _start_column, included_ranges) = blockquote_injection_data(text);
+        assert!(
+            included_ranges.is_some(),
+            "Blockquote should have included ranges"
+        );
+
+        let clean = extract_clean_content(text, byte_range, included_ranges.as_deref());
+        assert!(
+            !clean.contains("> "),
+            "Clean content should not contain blockquote prefixes: {:?}",
+            clean
+        );
+        assert!(
+            clean.contains("local x = 1"),
+            "Clean content should contain the code: {:?}",
+            clean
+        );
+        assert!(
+            clean.contains("local y = 2"),
+            "Clean content should contain the code: {:?}",
+            clean
+        );
+    }
+
+    #[test]
+    fn compute_line_column_offsets_without_ranges_returns_start_column() {
+        // When called with None for included_ranges, returns vec![start_column]
+        let offsets = compute_line_column_offsets("hello", 0..5, 4, None);
+        assert_eq!(offsets, vec![4]);
+    }
+
+    #[test]
+    fn compute_line_column_offsets_with_ranges_returns_per_line() {
+        // Blockquote case: "> " prefix on each line → 2 UTF-16 units per line
+        let text = "> ```lua\n> local x = 1\n> local y = 2\n> ```\n";
+        let (byte_range, start_column, included_ranges) = blockquote_injection_data(text);
+        assert!(included_ranges.is_some());
+
+        let offsets =
+            compute_line_column_offsets(text, byte_range, start_column, included_ranges.as_deref());
+        // Virtual lines 0 and 1 contain actual code after "> " prefixes.
+        // Line 2 may be trailing content (the "> " before "```") — its offset
+        // is 0 because there's no gap range starting on that line.
+        assert!(
+            offsets.len() >= 2,
+            "Should have at least 2 line offsets: {:?}",
+            offsets
+        );
+        assert_eq!(
+            offsets[0], 2,
+            "Line 0 should have offset 2 (for '> ' prefix)"
+        );
+        assert_eq!(
+            offsets[1], 2,
+            "Line 1 should have offset 2 (for '> ' prefix)"
+        );
+    }
+
+    /// Tree-sitter assigns column positions from Range.start_point, not byte offset.
+    ///
+    /// When `set_included_ranges` is called with ranges whose `start_point.column = 2`,
+    /// tree-sitter reports parsed nodes at column 2 — not column 0 — even though the
+    /// bytes start at offset 2 within the raw text.
+    ///
+    /// This is the invariant that makes blockquote injection work correctly:
+    /// `compute_included_ranges` builds ranges with `start_point.column = prefix_len`
+    /// (e.g. 2 for `> `), so injected keywords appear at their true host column.
+    #[test]
+    fn test_parse_with_included_ranges_preserves_start_point_column() {
+        // Simulate "> let x = 1;\n> let y = 2;\n" with included_ranges skipping `> `
+        // Line 0: "> let x = 1;\n" = 13 bytes (0..13, \n at byte 12)
+        // Line 1: "> let y = 2;\n" = 13 bytes (13..26, \n at byte 25)
+        // Content ranges (skipping 2-byte `> ` prefix on each line):
+        //   Range 0: bytes  2..13 = "let x = 1;\n", start_point {row:0, col:2}
+        //   Range 1: bytes 15..26 = "let y = 2;\n", start_point {row:1, col:2}
+        let content_text = "> let x = 1;\n> let y = 2;\n";
+        let included_ranges = vec![
+            tree_sitter::Range {
+                start_byte: 2,
+                end_byte: 13,
+                start_point: tree_sitter::Point { row: 0, column: 2 },
+                end_point: tree_sitter::Point { row: 1, column: 0 },
+            },
+            tree_sitter::Range {
+                start_byte: 15,
+                end_byte: 26,
+                start_point: tree_sitter::Point { row: 1, column: 2 },
+                end_point: tree_sitter::Point { row: 2, column: 0 },
+            },
+        ];
+
+        let rust_language: tree_sitter::Language = tree_sitter_rust::LANGUAGE.into();
+        let mut parser = Parser::new();
+        parser.set_language(&rust_language).unwrap();
+        let tree = parse_with_ranges(
+            &mut parser,
+            content_text,
+            Some(&included_ranges),
+            "test",
+            "rust",
+        )
+        .expect("should parse");
+
+        let query = Query::new(&rust_language, "(let_declaration) @decl").unwrap();
+        let mut cursor = tree_sitter::QueryCursor::new();
+        let mut matches = cursor.matches(&query, tree.root_node(), content_text.as_bytes());
+
+        let mut let_columns: Vec<usize> = Vec::new();
+        while let Some(m) = matches.next() {
+            for c in m.captures {
+                let node = c.node;
+                let mut walk = node.walk();
+                for child in node.children(&mut walk) {
+                    if child.kind() == "let" {
+                        let_columns.push(child.start_position().column);
+                    }
+                }
+            }
+        }
+
+        assert_eq!(
+            let_columns,
+            vec![2, 2],
+            "Both 'let' keywords should be at column 2 (from Range.start_point), not 0"
+        );
+    }
+}
